Solve LifeCycle
